Format code.
